backend/ggok/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.3
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
django-rest-swagger==2.2.0
djangorestframework==3.15.1
drf-yasg==1.21.7
gunicorn==22.0.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
Markdown==3.6
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==24.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
setuptools==69.2.0
simplejson==3.19.2
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
pillow==10.3.0
six==1.16.0
boto3==1.34.111
botocore==1.34.111
django-storages==1.14.3
